Batch claiming in download.py
